umi: Create a fake prebuilt_hidl_interfaces module for soong
Qualcomm has merged support for prebuilt_hidl_interfaces from master/keystone and started using it. AOSPA team picked it up and now they use it for their common components' init module. Thankfully we don't use that one (as we are "vendorless") so just make a dummy module type to allow it to build on aosp-based roms.

Sidenote: go sucks

Signed-off-by: Kuba Wojciechowski <nullbytepl@gmail.com>